Completely refactor the transport layer.
This change does refactor the client transport layer in order to make
room for a Hono client implementation. And others as well.

Before the change the AMQP transport does a few assumptions which aren't
valid with Hono. Also didn't the transport API allow for much variety
on the request level.

Now it is possible to assemble most parts of the request in a fluent way
and then execute the prepared request. For this the transport layer was
split off to allow a simple "Proton Message" based request/response
pattern, without any assumptions on the payload or the disposition
handling.

The next layer is a fluent API layer to create request/response patterns
and allow to chain in response handlers.

The final layer is the existing (for Hono new) layer of project/client
specific specializations.

I am aware of the fact that the middle layer looks rather complex. And
I'm still looking for ways to simplify it without loosing the
abstractions. Maybe we don't need some of the abstractions. But the
client specific layer (which you should only write once) is rather slim
and reduced complexity again by specializing on the interfaces and
superclasses.